\name{HW2Optim-package}
\alias{HW2Optim-package}
\alias{HW2Optim}
\docType{package}
\title{
\packageTitle{HW2Optim}
}
\description{
\packageDescription{HW2Optim}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HW2Optim}
\packageIndices{HW2Optim}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{HW2Optim}

Maintainer: \packageMaintainer{HW2Optim}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
